config enabled - check if config.ini already exists

+++ src/toHTML.py
@@ -23,31 +23,37 @@ class QConfig:
             #Check if the tmp folder  already exists
             if not os.path.exists(target):
                 os.makedirs(target)
+            else:
+                print('\tINFO: {} already exists so nothing was done...'.format(directory))
 
     def __createConfigFile(self):
         #Get path components of the current path
         self.__initPaths()
-        #Instantiate a config parser
-        config = configparser.ConfigParser()
-        #Intialize key paths
-        img_folder = os.path.join(utils.getPath(self.rootPath), QConfig.__directories[0])
-        data_folder = os.path.join(utils.getPath(self.rootPath), QConfig.__directories[1])
-        tmp_folder = os.path.join(utils.getPath(self.rootPath), QConfig.__directories[2])
-        report_folder = os.path.join(utils.getPath(self.rootPath), QConfig.__directories[3])
-        template_folder = os.path.join(utils.getPath(self.rootPath), QConfig.__directories[4])
+        #Create a config.ini file if one does not exist
+        if not os.path.exists('config.ini'):
+            #Instantiate a config parser
+            config = configparser.ConfigParser()
+            #Intialize key paths
+            img_folder = os.path.join(utils.getPath(self.rootPath), QConfig.__directories[0])
+            data_folder = os.path.join(utils.getPath(self.rootPath), QConfig.__directories[1])
+            tmp_folder = os.path.join(utils.getPath(self.rootPath), QConfig.__directories[2])
+            report_folder = os.path.join(utils.getPath(self.rootPath), QConfig.__directories[3])
+            template_folder = os.path.join(utils.getPath(self.rootPath), QConfig.__directories[4])
 
-        #Define sections in the config file
-        config['Locations'] = {'imageLocation': img_folder,
+            #Define sections in the config file
+            config['Locations'] = {'imageLocation': img_folder,
                              'dataLocation':  data_folder,
                              'tempLocation':  tmp_folder,
                              'reportLocation': report_folder,
                              'templateLocation': template_folder,
                             }
-        config['DataFrame']    = {'indexColumn':'LightID'}
-        config['ReportFormat'] = {'numberOfRowsPerPage':15}
+            config['DataFrame']    = {'indexColumn':'LightID'}
+            config['ReportFormat'] = {'numberOfRowsPerPage':15}
 
-        with open('config.ini', 'w') as config_file:
-            config.write(config_file)
+            with open('config.ini', 'w') as config_file:
+                config.write(config_file)
+        else:
+            print('\tINFO: config.ini already exists so nothing was done...')
 
     def __init__(self):
         #Initialize config file
@@ -128,7 +134,6 @@ if __name__ == '__main__':
                 'agent_name':'FBT_Sp'
                }
 
-    
     report = QReport( 'C:\Workspace\Quintus\data\m_data.csv' , **metaData )
     report.toHTML()
